hardware: blacklist nouveau & enable finegrain PM
The system frequently freezes on rebuilds, shutdowns and waking up from
sleep. This also happens randomly sometimes, but not too frequently.

This is currently an issue with the Nvidia 550 drivers, so I'm testing
if disabling nouveau and enabling finegrain power management can
mitigate it. Else, I will have to roll back the drivers to an earlier
version (possibly 535).

More details:
https://forums.developer.nvidia.com/t/series-550-freezes-laptop/284772/1