fixup: don't force X11 for QT apps
This was originally to fix some QT apps that have missing dock icons
when running under Wayland. However, this caused an issue with Calibre,
when PaperWM is enabled, making calibre-reader not render when switching
to another screen.

I can just not force XWayland for Calibre and still have the icons,
which I might do in the future if the icon issue persists or annoys me.